pyproject.toml
[project]
name = "complianceqapipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "azure-identity>=1.25.1",
    "azure-monitor-opentelemetry>=1.8.2",
    "azure-search-documents>=11.6.0",
    "azure-storage-blob>=12.28.0",
    "fastapi>=0.128.5",
    "firecrawl-py>=4.14.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.11",
    "langsmith>=0.4.37",
    "opentelemetry-instrumentation-fastapi>=0.60b0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "streamlit>=1.50.0",
    "uvicorn>=0.39.0",
    "yt-dlp>=2025.10.14",
]
